Update whois privacy related function APIs
To be the following:
- Dnsimple.Registrar.whois_privacy
- Dnsimple.Registrar.enable_whois_privacy
- Dnsimple.Registrar.disable_whois_privacy

These functions will be under the same module as the existing registrar
related ones so we expose the same API as in other clients.
